requirements.txt
aiogram==2.23.1
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.1.0
autopep8==2.0.1
Babel==2.9.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.4
frozenlist==1.3.3
future==0.18.2
idna==3.4
magic-filter==1.0.9
multidict==6.0.3
pycodestyle==2.10.0
pycparser==2.21
python-dotenv==0.14.0
pytz==2022.7
redis==3.5.3
requests==2.28.1
six==1.16.0
tomli==2.0.1
urllib3==1.26.13
yarl==1.8.2
